requirements.txt
bcrypt==4.0.1
fastapi==0.88.0
httpx==0.23.1
pydantic==1.10.2
PyJWT==2.6.0
PyMySQL==1.0.2
pytest==7.2.0
python-dotenv==0.21.0
SQLAlchemy==1.4.45
starlette==0.22.0
uvicorn==0.20.0